+++ readmeai/main.py
@@ -1,15 +1,17 @@
 #!/usr/bin/env python3
 
-import argparse
 import asyncio
 import os
-from typing import Dict, List, Tuple
+from typing import Dict, List, Optional, Tuple
 
-import conf
-import preprocess
-from builder import build_markdown_file
-from logger import Logger
-from model import OpenAIHandler
+import click
+
+from . import builder, conf, logger, model, preprocess
+
+logger = logger.Logger(__name__)
+config = conf.load_config()
+config_model = conf.AppConfigModel(app=config)
+config_helper = conf.load_config_helper(config_model)
 
 
 async def main(api_key: str, output: str, repository: str) -> None:
@@ -18,22 +20,20 @@ async def main(api_key: str, output: str, repository: str) -> None:
     conf.GitConfig.validate_repository(repository)
     config.git = conf.GitConfig(repository=repository)
     config.paths.readme = output
-
-    LOGGER.info("Model: %s", dict(config.api, api_key=("*" * 16)))
-    LOGGER.info("Repository: %s", config.git)
-
-    llm = OpenAIHandler(config)
+    logger.info("Model: %s", dict(config.api, api_key=("*" * 16)))
+    logger.info("Repository: %s", config.git)
+    llm = model.OpenAIHandler(config)
     await generate_readme(llm)
 
 
-async def generate_readme(llm: OpenAIHandler) -> None:
+async def generate_readme(llm: model.OpenAIHandler) -> None:
     """Orchestrates the README file generation process."""
     name, repository = config.git.name, config.git.repository
     scanner = preprocess.RepositoryParserWrapper(config, config_helper)
     dependencies, file_text = get_dependencies(scanner, repository)
 
-    LOGGER.info(f"Dependencies: {dependencies}")
-    LOGGER.info(f"Total files: {len(file_text)}")
+    logger.info(f"Dependencies: {dependencies}")
+    logger.info(f"Total files: {len(file_text)}")
 
     code_summary, slogan, overview, features = {}, "", "", ""
     try:
@@ -42,12 +42,12 @@ async def generate_readme(llm: OpenAIHandler) -> None:
             llm, repository, code_summary
         )
     except Exception as excinfo:
-        LOGGER.error(f"Exception: {excinfo}")
+        logger.error(f"Exception: {excinfo}")
     finally:
         await llm.close()
 
     format_markdown_sections(name, slogan, overview, features)
-    build_markdown_file(config, config_helper, dependencies, code_summary)
+    builder.build_markdown_file(config, config_helper, dependencies, code_summary)
 
 
 def format_markdown_sections(
@@ -58,7 +58,9 @@ def format_markdown_sections(
     config.md.intro = config.md.intro.format(overview, features)
 
 
-async def generate_code_to_text(llm: OpenAIHandler, file_text: str) -> Dict[str, str]:
+async def generate_code_to_text(
+    llm: model.OpenAIHandler, file_text: str
+) -> Dict[str, str]:
     """Generates code_to_text using llm."""
     return await llm.code_to_text(
         config_helper.ignore_files, file_text, config.prompts.code_summary
@@ -66,7 +68,7 @@ async def generate_code_to_text(llm: OpenAIHandler, file_text: str) -> Dict[str,
 
 
 async def generate_markdown_text(
-    llm: OpenAIHandler, repository: str, code_summary: str
+    llm: model.OpenAIHandler, repository: str, code_summary: str
 ) -> Tuple[str, str, str]:
     """Generates slogan, overview and features using llm."""
     prompts = [
@@ -86,45 +88,47 @@ def get_dependencies(
     return dependencies, file_text
 
 
+@click.command()
+@click.option(
+    "-k",
+    "--api-key",
+    default=os.environ.get("OPENAI_API_KEY", ""),
+    help="OpenAI API secret key.",
+)
+@click.option(
+    "-o",
+    "--output",
+    default="readme-ai.md",
+    help="README.md output file path.",
+)
+@click.option(
+    "-r",
+    "--repository",
+    required=True,
+    help="Repository URL or directory path.",
+)
+@click.option(
+    "-t",
+    "--template",
+    help="Template to use for README.md file.",
+)
+@click.option(
+    "-l",
+    "--language",
+    help="Language to write README.md file in.",
+)
+def cli(
+    api_key: str,
+    output: str,
+    repository: str,
+    template: Optional[int],
+    language: Optional[str],
+) -> None:
+    """Cli entrypoint for readme-ai pypi package."""
+    logger.info("README-AI is now executing.")
+    asyncio.run(main(api_key, output, repository))
+    logger.info("README-AI execution complete.\n")
+
+
 if __name__ == "__main__":
-    LOGGER = Logger(__name__)
-    LOGGER.info("README-AI is now executing.")
-
-    parser = argparse.ArgumentParser(
-        description="Generates a README.md file for a given repository."
-    )
-    parser.add_argument(
-        "-k",
-        "--api-key",
-        default=os.environ["OPENAI_API_KEY"],
-        help="OpenAI API secret key.",
-    )
-    parser.add_argument(
-        "-o",
-        "--output",
-        default="readme-ai.md",
-        help="README.md output file path.",
-    )
-    parser.add_argument(
-        "-r",
-        "--repository",
-        help="Repository URL or directory path.",
-    )
-    parser.add_argument(
-        "-t",
-        "--template",
-        help="Template to use for README.md file.",
-    )
-    parser.add_argument(
-        "-l",
-        "--language",
-        help="Language to write README.md file in.",
-    )
-    args = parser.parse_args()
-    config = conf.load_config()
-    config_model = conf.AppConfigModel(app=config)
-    config_helper = conf.load_config_helper(config_model)
-
-    asyncio.run(main(args.api_key, args.output, args.repository))
-
-    LOGGER.info("README-AI execution complete.\n")
+    cli()