+++ setup/setup.sh
@@ -61,12 +61,12 @@ if ! command -v conda &> /dev/null; then
     exit 1
 fi
 
-# Check for Python installation and required version (3.7+)
-echo "Checking for Python 3.7 or higher..."
+# Check for Python installation and required version (3.8+)
+echo "Checking for Python 3.8 or higher..."
 python_version=$(python3 -c 'import sys; print("{}.{}".format(sys.version_info[0], sys.version_info[1]))' 2>/dev/null)
 
 if [ $? -ne 0 ]; then
-    echo "Python 3 is not installed. Please install Python 3.7 or higher and rerun the script."
+    echo "Python 3 is not installed. Please install Python 3.8 or higher and rerun the script."
     exit 1
 fi
 
@@ -75,7 +75,7 @@ major=${version_parts[0]}
 minor=${version_parts[1]}
 
 if [ "$major" -lt 3 ] || ([ "$major" -eq 3 ] && [ "$minor" -lt 7 ]); then
-    echo "Please upgrade to Python 3.7 or higher and rerun the script."
+    echo "Please upgrade to Python 3.8 or higher and rerun the script."
     exit 1
 fi